+++ docs/testable.md
@@ -1,0 +1,3 @@
+# Первая проверка тестируемости
+ExitCode.kt, Action.kt - не требует тестирования (это просто enum)
+IAuthService, IPermissionManager, IResource - интерфейсы, не тестируются.
