+++ src/models/Resource.kt
@@ -1,5 +1,6 @@
 package models
 import interfaces.IResource
+import enumerators.Action
 class Resource (
     val name: String,
     val maxVolume: Int = 10,
